Add .gitignore, .idea folder, imports.razor file, app.css, and bootstrap.min.css
This commit includes the addition of several files including .gitignore and app.css. The .gitignore file is used to define the list of files and directories that should be ignored by git. The app.css file contains the main CSS styles for the application. Bootstrap.min.css has also been added as part of the essential styling files for Bootstrap, a powerful framework for building responsive webpages. The _Imports.razor file and .idea folder have also been created as part of the initial project setup.