Add PDF viewer, implement notes page, and remove counter page
This commit introduces a PDF viewer feature and adds a new Notes page. It also removes the previously used Counter page, alongside minor corrections to other pages to improve the overall functionality of the site. Syntax cleaning was done in CourseSection.cs for better readability and SyncFusion Blazor package was added to support implementation of the PDF viewer.