Add "Apuntes" page and update JavaScript library mappings
A new "Apuntes" (Notes) page was added with a PDF layout under the Memes2 section. The navigation menu has been updated to include a direct link to this new page. Additionally, the JavaScript library mappings for the project were updated. A minor update was also made in the Program.cs file replacing service scope factory with an underscore for unused parameter.